[scheduler] Move the form selectors to their own file
Follows the scheduler convention of keeping selector objects out of the
store class file.
